+++ tests/ReportServiceTest.php
@@ -162,6 +162,31 @@ class ReportServiceTest extends TestCase
         $filesystem->remove($reportDirectory);
     }
 
+    public function testCanGenerateCoberturaReport(): void
+    {
+        $filesystem = new Filesystem();
+        $reportFilename = sys_get_temp_dir() . '/cobertura.xml';
+        $filesystem->remove($reportFilename);
+
+        $driver = $this->createMock(Driver::class);
+        $coverage = new CodeCoverage($driver, new Filter());
+
+        $reportService = new ReportService(
+            [
+                'cobertura' => [
+                    'target' => $reportFilename,
+                ],
+            ]
+        );
+
+        $reportService->generateReport($coverage);
+        $report = file_get_contents($reportFilename);
+
+        self::assertNotEmpty($report);
+
+        $filesystem->remove($reportFilename);
+    }
+
     public function testCanGenerateMultipleReport(): void
     {
         $filesystem = new Filesystem();